Remove old dependency_finder and substitute with new version
